refactor: internal_infrastructureパッケージのバレルファイルを削除
- internal_infrastructure.dartバレルファイルを削除
- src/operational_settingsをlib/operational_settingsに移動
- src/shared_preferencesをlib/shared_preferencesに移動
- src/theme_settingをlib/theme_settingに移動
- 不要なinternal_infrastructure_base.dartとinternal_infrastructure_test.dartを削除
- 依存ファイルのimport文を更新
