+++ apps/app/lib/composition_root/data_sources/shared_preference_data_source.dart
@@ -1,5 +1,5 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:internal_infrastructure/internal_infrastructure.dart';
+import 'package:internal_infrastructure/shared_preferences/shared_preference_data_source.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 import 'package:shared_preferences/shared_preferences.dart';
