+++ apps/app/lib/composition_root/repositories/operational_settings_repository.dart
@@ -1,6 +1,6 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:internal_domain_logic/internal_domain_logic.dart';
-import 'package:internal_infrastructure/internal_infrastructure.dart';
+import 'package:internal_infrastructure/operational_settings/operational_settings_repository_impl.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
 part 'operational_settings_repository.g.dart';
