--- packages/infrastructure/lib/internal_infrastructure.dart
@@ -1,16 +1,0 @@
-/// Infrastructure layer package for data access
-/// and external service integration.
-///
-/// This package contains:
-/// - Repository implementations
-/// - API clients
-/// - Data storage services
-library;
-
-export 'src/internal_infrastructure_base.dart';
-export 'src/operational_settings/operational_settings_repository_impl.dart';
-export 'src/shared_preferences/shared_preference_data_source.dart';
-export 'src/shared_preferences/shared_preferences_keys.dart';
-export 'src/theme_setting/theme_setting_repository_impl.dart';
-
-// TODO: Export any libraries intended for clients of this package.
