+++ packages/infrastructure/lib/theme_setting/theme_setting_repository_impl.dart
@@ -1,6 +1,6 @@
 import 'package:internal_domain_logic/internal_domain_logic.dart';
 import 'package:internal_domain_model/internal_domain_model.dart';
-import 'package:internal_infrastructure/src/shared_preferences/shared_preference_data_source.dart';
+import 'package:internal_infrastructure/shared_preferences/shared_preference_data_source.dart';
 
 class ThemeSettingRepositoryImpl implements ThemeSettingRepository {
   const ThemeSettingRepositoryImpl(this._sharedPreference);
