--- packages/infrastructure/test/internal_infrastructure_test.dart
@@ -1,16 +1,0 @@
-import 'package:flutter_test/flutter_test.dart';
-import 'package:internal_infrastructure/internal_infrastructure.dart';
-
-void main() {
-  group('A group of tests', () {
-    final awesome = Awesome();
-
-    setUp(() {
-      // Additional setup goes here.
-    });
-
-    test('First Test', () {
-      expect(awesome.isAwesome, isTrue);
-    });
-  });
-}
